Move `GetImprudences` message into a new sub module
